refactor: move functions to utils.py
